Cleanup AutocompleteCategory.vue
